Rename markor baseactivity and use everywhere, move OpenEditor activity to subdir, by @gsantner
