+++ app/src/main/java/net/gsantner/markor/util/ShortcutUtils.java
@@ -19,7 +19,7 @@ import android.os.Build;
 import android.support.annotation.NonNull;
 
 import net.gsantner.markor.R;
-import net.gsantner.markor.activity.DocumentRelayActivity;
+import net.gsantner.markor.activity.openeditor.OpenEditorFromShortcutOrWidgetActivity;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -61,7 +61,7 @@ public class ShortcutUtils {
             final AppSettings appSettings = new AppSettings(context);
 
             // Create the to-do shortcut
-            Intent openTodo = new Intent(context, DocumentRelayActivity.class)
+            Intent openTodo = new Intent(context, OpenEditorFromShortcutOrWidgetActivity.class)
                     .setAction(Intent.ACTION_EDIT)
                     .setData(Uri.fromFile(appSettings.getTodoFile()));
 
@@ -74,7 +74,7 @@ public class ShortcutUtils {
             newShortcuts.add(shortcutToDo);
 
             // Create the QuickNote shortcut
-            Intent openQuickNote = new Intent(context, DocumentRelayActivity.class)
+            Intent openQuickNote = new Intent(context, OpenEditorFromShortcutOrWidgetActivity.class)
                     .setAction(Intent.ACTION_EDIT)
                     .setData(Uri.fromFile(appSettings.getQuickNoteFile()));
 
@@ -97,7 +97,7 @@ public class ShortcutUtils {
 
                 File file = new File(filePath);
 
-                Intent openFile = new Intent(context, DocumentRelayActivity.class)
+                Intent openFile = new Intent(context, OpenEditorFromShortcutOrWidgetActivity.class)
                         .setAction(Intent.ACTION_EDIT)
                         .setData(Uri.fromFile(file));
